refactor: moves components theme interfaces from Flowbyte.ts to their main files
