+++ src/lib/components/Accordion/Accordion.tsx
@@ -2,11 +2,33 @@ import classNames from 'classnames';
 import type { ComponentProps, FC, PropsWithChildren, ReactElement } from 'react';
 import { Children, cloneElement, useMemo, useState } from 'react';
 import { HiChevronDown } from 'react-icons/hi';
+import { FlowbiteBoolean } from '../Flowbite/FlowbiteTheme';
 import { useTheme } from '../Flowbite/ThemeContext';
 import { AccordionContent } from './AccordionContent';
 import type { AccordionPanelProps } from './AccordionPanel';
 import { AccordionPanel } from './AccordionPanel';
 import { AccordionTitle } from './AccordionTitle';
+
+export interface FlowbiteAccordionTheme {
+  base: string;
+  content: {
+    base: string;
+  };
+  flush: FlowbiteBoolean;
+  title: {
+    arrow: {
+      base: string;
+      open: {
+        off: string;
+        on: string;
+      };
+    };
+    base: string;
+    flush: FlowbiteBoolean;
+    heading: string;
+    open: FlowbiteBoolean;
+  };
+}
 
 export interface AccordionProps extends PropsWithChildren<ComponentProps<'div'>> {
   alwaysOpen?: boolean;
