+++ src/lib/components/Avatar/AvatarGroup.tsx
@@ -3,6 +3,10 @@ import type { ComponentProps, PropsWithChildren } from 'react';
 import React from 'react';
 import { useTheme } from '../Flowbite';
 
+export interface FlowbiteAvatarGroupTheme {
+  base: string;
+}
+
 export type AvatarGroupProps = PropsWithChildren<ComponentProps<'div'>>;
 
 const AvatarGroup: React.FC<AvatarGroupProps> = ({ children, className }) => {
