+++ src/lib/components/Avatar/index.ts
@@ -1,3 +1,3 @@
 export * from './Avatar';
-export type { AvatarGroupProps } from './AvatarGroup';
-export type { AvatarGroupdCounterProps } from './AvatarGroupCounter';
+export type { AvatarGroupProps, FlowbiteAvatarGroupTheme } from './AvatarGroup';
+export type { AvatarGroupdCounterProps, FlowbiteAvatarGroupCounterTheme } from './AvatarGroupCounter';
