+++ src/lib/components/Dropdown/Dropdown.tsx
@@ -4,12 +4,28 @@ import { HiOutlineChevronDown, HiOutlineChevronLeft, HiOutlineChevronRight, HiOu
 import { uuid } from '../../helpers/uuid';
 import type { ButtonProps } from '../Button';
 import { Button } from '../Button';
-import type { FloatingProps } from '../Floating';
+import type { FloatingProps, FlowbiteFloatingTheme } from '../Floating';
 import { Floating } from '../Floating';
 import { useTheme } from '../Flowbite/ThemeContext';
 import { DropdownDivider } from './DropdownDivider';
 import { DropdownHeader } from './DropdownHeader';
 import { DropdownItem } from './DropdownItem';
+
+export interface FlowbiteDropdownFloatingTheme extends FlowbiteFloatingTheme {
+  header: string;
+  item: {
+    base: string;
+    icon: string;
+  };
+  divider: string;
+}
+
+export interface FlowbiteDropdownTheme {
+  floating: FlowbiteDropdownFloatingTheme;
+  content: string;
+  inlineWrapper: string;
+  arrowIcon: string;
+}
 
 export interface DropdownProps extends PropsWithChildren<Pick<FloatingProps, 'placement' | 'trigger'>>, ButtonProps {
   label: ReactNode;
