+++ src/lib/components/Flowbite/FlowbiteTheme.ts
@@ -1,30 +1,27 @@
 import { DeepPartial } from '..';
-import type { AlertColors } from '../Alert/Alert';
-import type { AvatarColors, AvatarSizes } from '../Avatar/Avatar';
-import type { BadgeColors, BadgeSizes } from '../Badge';
-import type {
-  ButtonColors,
-  ButtonGradientColors,
-  ButtonGradientDuoToneColors,
-  ButtonOutlineColors,
-  ButtonSizes,
-} from '../Button';
-import { PositionInButtonGroup } from '../Button';
-import type {
-  HelperColors,
-  LabelColors,
-  SelectColors,
-  SelectSizes,
-  TextareaColors,
-  TextInputColors,
-  TextInputSizes,
-} from '../FormControls';
-import type { ModalPositions, ModalSizes } from '../Modal';
-import type { ProgressColor, ProgressSizes } from '../Progress';
-import type { StarSizes } from '../Rating';
-import type { SidebarCTAColors } from '../Sidebar/SidebarCTA';
-import type { SpinnerColors, SpinnerSizes } from '../Spinner/Spinner';
-import type { TabStyleItem, TabStyles } from '../Tab/Tabs';
+import { FlowbiteAccordionTheme } from '../Accordion';
+import { FlowbiteAlertTheme } from '../Alert';
+import { FlowbiteAvatarGroupCounterTheme, FlowbiteAvatarGroupTheme, FlowbiteAvatarTheme } from '../Avatar';
+import { FlowbiteBadgeTheme } from '../Badge';
+import { FlowbiteBreadcrumbTheme } from '../Breadcrumb';
+import { FlowbiteButtonGroupTheme, FlowbiteButtonTheme } from '../Button';
+import { FlowbiteCardTheme } from '../Card';
+import { FlowbiteCarouselTheme } from '../Carousel';
+import { FlowbiteDarkThemeToggleTheme } from '../DarkThemeToggle';
+import { FlowbiteDropdownTheme } from '../Dropdown';
+import { FlowbiteFooterTheme } from '../Footer';
+import { FlowbiteFormControlsTheme } from '../FormControls';
+import { FlowbiteListGroupTheme } from '../ListGroup';
+import { FlowbiteModalTheme } from '../Modal';
+import { FlowbiteNavbarTheme } from '../Navbar';
+import { FlowbitePaginationTheme } from '../Pagination';
+import { FlowbiteProgressTheme } from '../Progress';
+import { FlowbiteRatingTheme } from '../Rating';
+import { FlowbiteSidebarTheme } from '../Sidebar';
+import { FlowbiteSpinnerTheme } from '../Spinner';
+import { FlowbiteTabTheme } from '../Tab';
+import { FlowbiteToastTheme } from '../Toast';
+import { FlowbiteTooltipTheme } from '../Tooltip';
 
 export type CustomFlowbiteTheme = DeepPartial<FlowbiteTheme>;
 
@@ -55,531 +52,6 @@ export interface FlowbiteTheme extends Record<string, unknown> {
   toast: FlowbiteToastTheme;
   tooltip: FlowbiteTooltipTheme;
   dropdown: FlowbiteDropdownTheme;
-}
-
-export interface FlowbiteAccordionTheme {
-  base: string;
-  content: {
-    base: string;
-  };
-  flush: FlowbiteBoolean;
-  title: {
-    arrow: {
-      base: string;
-      open: {
-        off: string;
-        on: string;
-      };
-    };
-    base: string;
-    flush: FlowbiteBoolean;
-    heading: string;
-    open: FlowbiteBoolean;
-  };
-}
-
-export interface FlowbiteAlertTheme {
-  base: string;
-  borderAccent: string;
-  wrapper: string;
-  closeButton: {
-    base: string;
-    icon: string;
-    color: AlertColors;
-  };
-  color: AlertColors;
-  icon: string;
-  rounded: string;
-}
-
-export interface FlowbiteAvatarTheme {
-  base: string;
-  bordered: string;
-  img: {
-    off: string;
-    on: string;
-    placeholder: string;
-  };
-  color: AvatarColors;
-  rounded: string;
-  size: AvatarSizes;
-  stacked: string;
-  status: {
-    away: string;
-    base: string;
-    busy: string;
-    offline: string;
-    online: string;
-  };
-  statusPosition: FlowbitePositions;
-  initials: {
-    base: string;
-    text: string;
-  };
-}
-
-export interface FlowbiteAvatarGroupCounterTheme {
-  base: string;
-}
-
-export interface FlowbiteAvatarGroupTheme {
-  base: string;
-}
-
-export interface FlowbiteBadgeTheme {
-  base: string;
-  color: BadgeColors;
-  href: string;
-  icon: {
-    off: string;
-    on: string;
-    size: BadgeSizes;
-  };
-  size: BadgeSizes;
-}
-
-export interface FlowbiteBreadcrumbTheme {
-  item: {
-    base: string;
-    chevron: string;
-    href: {
-      off: string;
-      on: string;
-    };
-    icon: string;
-  };
-  list: string;
-}
-
-export interface FlowbiteButtonTheme {
-  base: string;
-  fullSized: string;
-  color: ButtonColors;
-  disabled: string;
-  gradient: ButtonGradientColors;
-  gradientDuoTone: ButtonGradientDuoToneColors;
-  inner: {
-    base: string;
-    position: PositionInButtonGroup;
-    outline: string;
-  };
-  label: string;
-  outline: FlowbiteBoolean & {
-    color: ButtonOutlineColors;
-    pill: FlowbiteBoolean;
-  };
-  pill: FlowbiteBoolean;
-  size: ButtonSizes;
-}
-
-export interface FlowbiteButtonGroupTheme {
-  base: string;
-  position: PositionInButtonGroup;
-}
-
-export interface FlowbiteCardTheme {
-  base: string;
-  children: string;
-  horizontal: {
-    off: string;
-    on: string;
-  };
-  href: string;
-  img: {
-    base: string;
-    horizontal: {
-      off: string;
-      on: string;
-    };
-  };
-}
-
-export interface FlowbiteCarouselTheme {
-  base: string;
-  indicators: {
-    active: {
-      off: string;
-      on: string;
-    };
-    base: string;
-    wrapper: string;
-  };
-  item: {
-    base: string;
-    wrapper: string;
-  };
-  control: {
-    base: string;
-    icon: string;
-  };
-  leftControl: string;
-  rightControl: string;
-  scrollContainer: {
-    base: string;
-    snap: string;
-  };
-}
-
-export interface FlowbiteDarkThemeToggleTheme {
-  base: string;
-  icon: string;
-}
-
-export interface FlowbiteFooterTheme {
-  base: string;
-  container: string;
-  bgDark: string;
-  groupLink: {
-    base: string;
-    link: {
-      base: string;
-      href: string;
-    };
-    col: string;
-  };
-  icon: {
-    base: string;
-    size: string;
-  };
-  title: {
-    base: string;
-  };
-  divider: {
-    base: string;
-  };
-  copyright: {
-    base: string;
-    href: string;
-    span: string;
-  };
-  brand: {
-    base: string;
-    img: string;
-    span: string;
-  };
-}
-
-export interface FlowbiteFormControlsTheme {
-  helperText: {
-    base: string;
-    colors: HelperColors;
-  };
-  label: {
-    base: string;
-    colors: LabelColors;
-    disabled: string;
-  };
-  radio: {
-    base: string;
-  };
-  checkbox: {
-    base: string;
-  };
-  textInput: {
-    base: string;
-    addon: string;
-    field: {
-      base: string;
-      icon: {
-        base: string;
-        svg: string;
-      };
-      input: {
-        base: string;
-        sizes: TextInputSizes;
-        colors: TextInputColors;
-        withIcon: FlowbiteBoolean;
-        withAddon: FlowbiteBoolean;
-        withShadow: FlowbiteBoolean;
-      };
-    };
-  };
-  fileInput: {
-    base: string;
-    field: {
-      base: string;
-      input: {
-        base: string;
-        sizes: TextInputSizes;
-        colors: TextInputColors;
-      };
-    };
-  };
-  toggleSwitch: {
-    base: string;
-    active: FlowbiteBoolean;
-    toggle: {
-      base: string;
-      checked: FlowbiteBoolean & {
-        color: FlowbiteColors;
-      };
-    };
-    label: string;
-  };
-  textarea: {
-    base: string;
-    colors: TextareaColors;
-    withShadow: FlowbiteBoolean;
-  };
-  select: {
-    base: string;
-    addon: string;
-    field: {
-      base: string;
-      icon: {
-        base: string;
-        svg: string;
-      };
-      select: {
-        base: string;
-        withIcon: FlowbiteBoolean;
-        withAddon: FlowbiteBoolean;
-        withShadow: FlowbiteBoolean;
-        sizes: SelectSizes;
-        colors: SelectColors;
-      };
-    };
-  };
-}
-
-export interface FlowbiteListGroupTheme {
-  base: string;
-  item: {
-    active: FlowbiteBoolean;
-    base: string;
-    href: FlowbiteBoolean;
-    icon: string;
-  };
-}
-
-export interface FlowbiteModalTheme {
-  base: string;
-  show: FlowbiteBoolean;
-  content: {
-    base: string;
-    inner: string;
-  };
-  body: {
-    base: string;
-    popup: string;
-  };
-  header: {
-    base: string;
-    popup: string;
-    title: string;
-    close: {
-      base: string;
-      icon: string;
-    };
-  };
-  footer: {
-    base: string;
-    popup: string;
-  };
-  sizes: ModalSizes;
-  positions: ModalPositions;
-}
-
-export interface FlowbiteNavbarTheme {
-  base: string;
-  rounded: FlowbiteBoolean;
-  bordered: FlowbiteBoolean;
-  inner: {
-    base: string;
-    fluid: FlowbiteBoolean;
-  };
-  brand: string;
-  collapse: {
-    base: string;
-    list: string;
-    hidden: FlowbiteBoolean;
-  };
-  link: {
-    base: string;
-    active: FlowbiteBoolean;
-    disabled: FlowbiteBoolean;
-  };
-  toggle: {
-    base: string;
-    icon: string;
-  };
-}
-
-export interface FlowbiteRatingTheme {
-  base: string;
-  star: {
-    sizes: StarSizes;
-    filled: string;
-    empty: string;
-  };
-  advanced: {
-    base: string;
-    label: string;
-    progress: {
-      base: string;
-      fill: string;
-      label: string;
-    };
-  };
-}
-
-export interface FlowbitePaginationTheme {
-  base: string;
-  layout: {
-    table: {
-      base: string;
-      span: string;
-    };
-  };
-  pages: {
-    base: string;
-    showIcon: string;
-    previous: {
-      base: string;
-      icon: string;
-    };
-    next: {
-      base: string;
-      icon: string;
-    };
-    selector: {
-      base: string;
-      active: string;
-    };
-  };
-}
-
-export interface FlowbiteSidebarTheme {
-  base: string;
-  collapsed: FlowbiteBoolean;
-  inner: string;
-  collapse: {
-    button: string;
-    icon: {
-      base: string;
-      open: FlowbiteBoolean;
-    };
-    label: {
-      base: string;
-      icon: string;
-    };
-    list: string;
-  };
-  cta: {
-    base: string;
-    color: SidebarCTAColors;
-  };
-  item: {
-    active: string;
-    base: string;
-    collapsed: {
-      insideCollapse: string;
-      noIcon: string;
-    };
-    content: {
-      base: string;
-    };
-    label: string;
-    icon: {
-      base: string;
-      active: string;
-    };
-  };
-  items: string;
-  itemGroup: string;
-  logo: {
-    base: string;
-    collapsed: FlowbiteBoolean;
-    img: string;
-  };
-}
-
-export interface FlowbiteProgressTheme {
-  base: string;
-  label: string;
-  bar: string;
-  color: ProgressColor;
-  size: ProgressSizes;
-}
-
-export interface FlowbiteSpinnerTheme {
-  base: string;
-  color: SpinnerColors;
-  light: {
-    off: {
-      base: string;
-      color: SpinnerColors;
-    };
-    on: {
-      base: string;
-      color: SpinnerColors;
-    };
-  };
-  size: SpinnerSizes;
-}
-
-export interface FlowbiteTabTheme {
-  base: string;
-  tablist: {
-    base: string;
-    styles: TabStyles;
-    tabitem: {
-      base: string;
-      styles: TabStyleItem<TabStyles>;
-      icon: string;
-    };
-  };
-  tabpanel: string;
-}
-
-export interface FlowbiteToastTheme {
-  base: string;
-  closed: string;
-  removed: string;
-  toggle: {
-    base: string;
-    icon: string;
-  };
-}
-
-export interface FlowbiteFloatingTheme {
-  target: string;
-  base: string;
-  animation: string;
-  hidden: string;
-  style: {
-    dark: string;
-    light: string;
-    auto: string;
-  };
-  content: string;
-  arrow: {
-    base: string;
-    style: {
-      dark: string;
-      light: string;
-      auto: string;
-    };
-    placement: string;
-  };
-}
-
-export interface FlowbiteTooltipTheme extends FlowbiteFloatingTheme {}
-
-export interface FlowbiteDropdownFloatingTheme extends FlowbiteFloatingTheme {
-  header: string;
-  item: {
-    base: string;
-    icon: string;
-  };
-  divider: string;
-}
-
-export interface FlowbiteDropdownTheme {
-  floating: FlowbiteDropdownFloatingTheme;
-  content: string;
-  inlineWrapper: string;
-  arrowIcon: string;
 }
 
 export interface FlowbiteBoolean {
