+++ src/lib/components/FormControls/HelperText.tsx
@@ -1,7 +1,94 @@
 import classNames from 'classnames';
 import type { ComponentProps, FC, PropsWithChildren } from 'react';
-import type { FlowbiteColors } from '../Flowbite/FlowbiteTheme';
+import type { FlowbiteBoolean, FlowbiteColors } from '../Flowbite/FlowbiteTheme';
 import { useTheme } from '../Flowbite/ThemeContext';
+import { LabelColors } from './Label';
+import { SelectColors, SelectSizes } from './Select';
+import { TextareaColors } from './Textarea';
+import { TextInputColors, TextInputSizes } from './TextInput';
+
+export interface FlowbiteFormControlsTheme {
+  helperText: {
+    base: string;
+    colors: HelperColors;
+  };
+  label: {
+    base: string;
+    colors: LabelColors;
+    disabled: string;
+  };
+  radio: {
+    base: string;
+  };
+  checkbox: {
+    base: string;
+  };
+  textInput: {
+    base: string;
+    addon: string;
+    field: {
+      base: string;
+      icon: {
+        base: string;
+        svg: string;
+      };
+      input: {
+        base: string;
+        sizes: TextInputSizes;
+        colors: TextInputColors;
+        withIcon: FlowbiteBoolean;
+        withAddon: FlowbiteBoolean;
+        withShadow: FlowbiteBoolean;
+      };
+    };
+  };
+  fileInput: {
+    base: string;
+    field: {
+      base: string;
+      input: {
+        base: string;
+        sizes: TextInputSizes;
+        colors: TextInputColors;
+      };
+    };
+  };
+  toggleSwitch: {
+    base: string;
+    active: FlowbiteBoolean;
+    toggle: {
+      base: string;
+      checked: FlowbiteBoolean & {
+        color: FlowbiteColors;
+      };
+    };
+    label: string;
+  };
+  textarea: {
+    base: string;
+    colors: TextareaColors;
+    withShadow: FlowbiteBoolean;
+  };
+  select: {
+    base: string;
+    addon: string;
+    field: {
+      base: string;
+      icon: {
+        base: string;
+        svg: string;
+      };
+      select: {
+        base: string;
+        withIcon: FlowbiteBoolean;
+        withAddon: FlowbiteBoolean;
+        withShadow: FlowbiteBoolean;
+        sizes: SelectSizes;
+        colors: SelectColors;
+      };
+    };
+  };
+}
 
 export interface HelperColors extends Pick<FlowbiteColors, 'gray' | 'info' | 'failure' | 'warning' | 'success'> {
   [key: string]: string;
