+++ src/lib/components/ListGroup/ListGroup.tsx
@@ -1,7 +1,18 @@
 import classNames from 'classnames';
 import type { ComponentProps, FC, PropsWithChildren } from 'react';
+import { FlowbiteBoolean } from '../Flowbite/FlowbiteTheme';
 import { useTheme } from '../Flowbite/ThemeContext';
 import { ListGroupItem } from './ListGroupItem';
+
+export interface FlowbiteListGroupTheme {
+  base: string;
+  item: {
+    active: FlowbiteBoolean;
+    base: string;
+    href: FlowbiteBoolean;
+    icon: string;
+  };
+}
 
 export type ListGroupProps = PropsWithChildren<ComponentProps<'div'>>;
