+++ src/lib/components/Rating/Rating.tsx
@@ -6,6 +6,24 @@ import { RatingAdvanced } from './RatingAdvanced';
 import { RatingContext } from './RatingContext';
 import { RatingStar } from './RatingStar';
 
+export interface FlowbiteRatingTheme {
+  base: string;
+  star: {
+    sizes: StarSizes;
+    filled: string;
+    empty: string;
+  };
+  advanced: {
+    base: string;
+    label: string;
+    progress: {
+      base: string;
+      fill: string;
+      label: string;
+    };
+  };
+}
+
 export interface StarSizes extends Pick<FlowbiteSizes, 'sm' | 'md' | 'lg'> {
   [key: string]: string;
 }
