+++ src/lib/components/Sidebar/Sidebar.tsx
@@ -1,13 +1,59 @@
 import classNames from 'classnames';
 import { ComponentProps, FC, PropsWithChildren } from 'react';
+import { FlowbiteBoolean } from '../Flowbite/FlowbiteTheme';
 import { useTheme } from '../Flowbite/ThemeContext';
 import SidebarCollapse from './SidebarCollapse';
 import { SidebarContext } from './SidebarContext';
-import SidebarCTA from './SidebarCTA';
+import SidebarCTA, { SidebarCTAColors } from './SidebarCTA';
 import SidebarItem from './SidebarItem';
 import SidebarItemGroup from './SidebarItemGroup';
 import SidebarItems from './SidebarItems';
 import SidebarLogo from './SidebarLogo';
+
+export interface FlowbiteSidebarTheme {
+  base: string;
+  collapsed: FlowbiteBoolean;
+  inner: string;
+  collapse: {
+    button: string;
+    icon: {
+      base: string;
+      open: FlowbiteBoolean;
+    };
+    label: {
+      base: string;
+      icon: string;
+    };
+    list: string;
+  };
+  cta: {
+    base: string;
+    color: SidebarCTAColors;
+  };
+  item: {
+    active: string;
+    base: string;
+    collapsed: {
+      insideCollapse: string;
+      noIcon: string;
+    };
+    content: {
+      base: string;
+    };
+    label: string;
+    icon: {
+      base: string;
+      active: string;
+    };
+  };
+  items: string;
+  itemGroup: string;
+  logo: {
+    base: string;
+    collapsed: FlowbiteBoolean;
+    img: string;
+  };
+}
 
 export interface SidebarProps extends PropsWithChildren<ComponentProps<'div'>> {
   collapseBehavior?: 'collapse' | 'hide';
